requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
groq==0.4.1
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
python-multipart==0.0.6
ffmpeg-python==0.2.0
openai-whisper==20231117
torch==2.0.1
torchaudio==2.0.2
sentence-transformers==2.2.2
llama-index==0.10.67
llama-index-vector-stores-faiss==0.1.2
llama-index-embeddings-huggingface==0.1.4
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
youtube-dl==2021.12.17
yt-dlp==2023.12.30
gunicorn==21.2.0 